+++ changelog.md
@@ -3,3 +3,5 @@
 ## 2.0.0
 * 1.20.1!
 * When out of lives, dying now instantly respawns you at the same location where you died
+* Players are now informed when max health is lost
+* Added a config option to cap the amount of max health modifier